Extract rest of the middlewares from prepare contex
