+++ lib/cartodb/controllers/analyses.js
@@ -1,11 +1,14 @@
 var PSQL = require('cartodb-psql');
 var cors = require('../middleware/cors');
 var userMiddleware = require('../middleware/user');
+const locals = require('../middleware/locals');
+const cleanUpQueryParams = require('../middleware/clean-up-query-params');
+const layergroupToken = require('../middleware/layergroup-token');
+const credentials = require('../middleware/credentials');
 const authorize = require('../middleware/authorize');
 const dbConnSetup = require('../middleware/db-conn-setup');
 
-function AnalysesController(prepareContext, pgConnection, authApi) {
-    this.prepareContext = prepareContext;
+function AnalysesController(pgConnection, authApi) {
     this.pgConnection = pgConnection;
     this.authApi = authApi;
 }
@@ -17,7 +20,10 @@ AnalysesController.prototype.register = function (app) {
         `${app.base_url_mapconfig}/analyses/catalog`,
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createPGClient(),
