+++ lib/cartodb/controllers/layergroup.js
@@ -2,6 +2,10 @@ const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
 const allowQueryParams = require('../middleware/allow-query-params');
 const vectorError = require('../middleware/vector-error');
+const locals = require('../middleware/locals');
+const cleanUpQueryParams = require('../middleware/clean-up-query-params');
+const layergroupToken = require('../middleware/layergroup-token');
+const credentials = require('../middleware/credentials');
 const dbConnSetup = require('../middleware/db-conn-setup');
 const authorize = require('../middleware/authorize');
 const DataviewBackend = require('../backends/dataview');
@@ -30,7 +34,7 @@ const SUPPORTED_FORMATS = {
  * @param {AnalysisBackend} analysisBackend
  * @constructor
  */
-function LayergroupController(prepareContext, pgConnection, mapStore, tileBackend, previewBackend, attributesBackend,
+function LayergroupController(pgConnection, mapStore, tileBackend, previewBackend, attributesBackend,
                               surrogateKeysCache, userLimitsApi, layergroupAffectedTables, analysisBackend, authApi) {
     this.pgConnection = pgConnection;
     this.mapStore = mapStore;
@@ -43,8 +47,6 @@ function LayergroupController(prepareContext, pgConnection, mapStore, tileBacken
 
     this.dataviewBackend = new DataviewBackend(analysisBackend);
     this.analysisStatusBackend = new AnalysisStatusBackend();
-
-    this.prepareContext = prepareContext;
     this.authApi = authApi;
 }
 
@@ -57,7 +59,10 @@ LayergroupController.prototype.register = function(app) {
         `${basePath}/:token/:z/:x/:y@:scale_factor?x.:format`,
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -78,7 +83,10 @@ LayergroupController.prototype.register = function(app) {
         `${basePath}/:token/:z/:x/:y.:format`,
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -100,7 +108,10 @@ LayergroupController.prototype.register = function(app) {
         distinguishLayergroupFromStaticRoute(),
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -121,7 +132,10 @@ LayergroupController.prototype.register = function(app) {
         `${basePath}/:token/:layer/attributes/:fid`,
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -141,7 +155,10 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(['layer']),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi, forcedFormat),
@@ -159,7 +176,10 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(['layer']),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi, forcedFormat),
@@ -195,7 +215,10 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -213,7 +236,10 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -231,7 +257,10 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -249,7 +278,10 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
@@ -266,7 +298,10 @@ LayergroupController.prototype.register = function(app) {
         `${basePath}/:token/analysis/node/:nodeId`,
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         analysisNodeStatus(this.analysisStatusBackend),
