+++ lib/cartodb/controllers/map.js
@@ -7,6 +7,10 @@ const ResourceLocator = require('../models/resource-locator');
 const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
 const allowQueryParams = require('../middleware/allow-query-params');
+const locals = require('../middleware/locals');
+const cleanUpQueryParams = require('../middleware/clean-up-query-params');
+const layergroupToken = require('../middleware/layergroup-token');
+const credentials = require('../middleware/credentials');
 const dbConnSetup = require('../middleware/db-conn-setup');
 const authorize = require('../middleware/authorize');
 const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
@@ -27,7 +31,7 @@ const LayergroupMetadata = require('../utils/layergroup-metadata');
  * @param {StatsBackend} statsBackend
  * @constructor
  */
-function MapController(prepareContext, pgConnection, templateMaps, mapBackend, metadataBackend,
+function MapController(pgConnection, templateMaps, mapBackend, metadataBackend,
                        surrogateKeysCache, userLimitsApi, layergroupAffectedTables, mapConfigAdapter,
                        statsBackend, authApi) {
     this.pgConnection = pgConnection;
@@ -43,7 +47,6 @@ function MapController(prepareContext, pgConnection, templateMaps, mapBackend, m
     this.layergroupMetadata = new LayergroupMetadata(resourceLocator);
 
     this.statsBackend = statsBackend;
-    this.prepareContext = prepareContext;
     this.authApi = authApi;
 }
 
@@ -71,7 +74,10 @@ MapController.prototype.composeCreateMapMiddleware = function (useTemplate = fal
         cors(),
         userMiddleware(),
         allowQueryParams(['aggregation']),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         initProfiler(isTemplateInstantiation),
