+++ lib/cartodb/controllers/named_maps.js
@@ -1,6 +1,10 @@
 const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
 const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
+const locals = require('../middleware/locals');
+const cleanUpQueryParams = require('../middleware/clean-up-query-params');
+const layergroupToken = require('../middleware/layergroup-token');
+const credentials = require('../middleware/credentials');
 const dbConnSetup = require('../middleware/db-conn-setup');
 const authorize = require('../middleware/authorize');
 const allowQueryParams = require('../middleware/allow-query-params');
@@ -29,7 +33,7 @@ function getRequestParams(locals) {
     return params;
 }
 
-function NamedMapsController(prepareContext, namedMapProviderCache, tileBackend, previewBackend,
+function NamedMapsController(namedMapProviderCache, tileBackend, previewBackend,
                              surrogateKeysCache, tablesExtentApi, metadataBackend, pgConnection, authApi) {
     this.namedMapProviderCache = namedMapProviderCache;
     this.tileBackend = tileBackend;
@@ -37,7 +41,6 @@ function NamedMapsController(prepareContext, namedMapProviderCache, tileBackend,
     this.surrogateKeysCache = surrogateKeysCache;
     this.tablesExtentApi = tablesExtentApi;
     this.metadataBackend = metadataBackend;
-    this.prepareContext = prepareContext;
     this.pgConnection = pgConnection;
     this.authApi = authApi;
 }
@@ -51,7 +54,10 @@ NamedMapsController.prototype.register = function(app) {
         `${base_url_templated}/:template_id/:layer/:z/:x/:y.(:format)`,
         cors(),
         userMiddleware(),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         getNamedMapProvider({
@@ -77,7 +83,10 @@ NamedMapsController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         allowQueryParams(['layer', 'zoom', 'lon', 'lat', 'bbox']),
-        this.prepareContext,
+        locals(),
+        cleanUpQueryParams(),
+        layergroupToken(),
+        credentials(),
         authorize(this.authApi),
         dbConnSetup(this.pgConnection),
         getNamedMapProvider({
