+++ lib/cartodb/controllers/named_maps_admin.js
@@ -1,8 +1,8 @@
 const { templateName } = require('../backends/template_maps');
 const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
-const localsMiddleware = require('../middleware/context/locals');
-const credentialsMiddleware = require('../middleware/context/credentials');
+const localsMiddleware = require('../middleware/locals');
+const credentialsMiddleware = require('../middleware/credentials');
 
 /**
  * @param {AuthApi} authApi
