+++ lib/cartodb/middleware/clean-up-query-params.js
