--- lib/cartodb/middleware/context/index.js
@@ -1,13 +1,0 @@
-const locals = require('./locals');
-const cleanUpQueryParams = require('./clean-up-query-params');
-const layergroupToken = require('./layergroup-token');
-const credentials = require('./credentials');
-
-module.exports = function prepareContextMiddleware() {
-    return [
-        locals(),
-        cleanUpQueryParams(),
-        layergroupToken(),
-        credentials(),
-    ];
-};
