+++ lib/cartodb/middleware/credentials.js
