+++ lib/cartodb/middleware/layergroup-token.js
@@ -1,4 +1,4 @@
-const LayergroupToken = require('../../models/layergroup-token');
+const LayergroupToken = require('../models/layergroup-token');
 const authErrorMessageTemplate = function (signer, user) {
     return `Cannot use map signature of user "${signer}" on db of user "${user}"`;
 };
