+++ lib/cartodb/middleware/locals.js
