+++ lib/cartodb/server.js
@@ -49,8 +49,6 @@ var StatsBackend = require('./backends/stats');
 const lzmaMiddleware = require('./middleware/lzma');
 const errorMiddleware = require('./middleware/error-middleware');
 
-const prepareContextMiddleware = require('./middleware/context');
-
 module.exports = function(serverOptions) {
     // Make stats client globally accessible
     global.statsClient = StatsClient.getInstance(serverOptions.statsd);
@@ -215,16 +213,11 @@ module.exports = function(serverOptions) {
 
     var versions = getAndValidateVersions(serverOptions);
 
-    const prepareContext = typeof serverOptions.req2params === 'function' ?
-        serverOptions.req2params :
-        prepareContextMiddleware();
-
     /*******************************************************************************************************************
      * Routing
      ******************************************************************************************************************/
 
     new controller.Layergroup(
-        prepareContext,
         pgConnection,
         mapStore,
         tileBackend,
@@ -238,7 +231,6 @@ module.exports = function(serverOptions) {
     ).register(app);
 
     new controller.Map(
-        prepareContext,
         pgConnection,
         templateMaps,
         mapBackend,
@@ -252,7 +244,6 @@ module.exports = function(serverOptions) {
     ).register(app);
 
     new controller.NamedMaps(
-        prepareContext,
         namedMapProviderCache,
         tileBackend,
         previewBackend,
@@ -265,7 +256,7 @@ module.exports = function(serverOptions) {
 
     new controller.NamedMapsAdmin(authApi, templateMaps).register(app);
 
-    new controller.Analyses(prepareContext, pgConnection, authApi).register(app);
+    new controller.Analyses(pgConnection, authApi).register(app);
 
     new controller.ServerInfo(versions).register(app);
